fix(warnings): drive compiler warnings to zero across both build configs
Full-clean census went from 492 unique warning sites (and ~640 counted
incrementally) to zero in both build-sanitized (hardening=basic) and
build (hardening=maximum, -Wformat=2). No pragmas added; every class
fixed at source.

Mechanical classes (~350 sites): -Wwritable-strings via const/const_cast
(repo convention), unused/unused-but-set deletions (dead state removed,
side-effecting calls kept), -Wswitch exhaustiveness, misleading-
indentation/dangling-else bracing, reorder-ctor, MSVC pragma guards.

Format safety (~250 sites):
- format(printf) attributes added to the printf wrappers (c3debug_dprintf,
  c3errors_ErrorDialog/FatalDialog, Log_Middle, ldlif_log, netconsole,
  c3files_fscanf/fprintf, DebugWindow::AddText, SetText2, SetControlText,
  combat_print, GameEventManager::Log) — this type-checks every DPRINTF
  call site for the first time and flushed out a decade of latent
  mismatches (%lx on uint32, Unit/Army ID handles, pointers, time_t,
  size_t), all fixed per established idioms (.m_id+%x, uintptr_t+%lx).
- Real bugs found and fixed by the new checking: format-string injection
  (computed text passed as format) in DBLexer, ColorSet, SlicFrame,
  sliccmd, sliciffile, SlicNamedSymbol (inverted module/fmt args),
  aui_sdlui (SDL_GetError), civapp, thronedb, c3errors internals,
  c3imageformats, ldlif, GameEventArgument, controlpanelwindow,
  net_action, net_info (arity bugs reading garbage from varargs),
  governor/scheduler (insufficient-args since the Activision import),
  network AddPlayer, TradeRouteData (missing valid arg), slicfuncai
  ($d typos).
- i18n templates: stringdb_FormatOr()/StringDB::GetTextOr() with
  format_arg fallback literals (gettext idiom) — localized templates
  flow into snprintf with call-site-checked fallbacks.
- spdlog/fmt/SDL deps now -isystem (include_type) so vendor headers
  stop firing -Wdeprecated-declarations through the shared
  /opt/homebrew/include search path.

Also: dead ErrMsg.cpp/ErrMsg.h deleted (zero users); switch-corruption
repairs (ArmyData UpdateZOCForMove braces, net_action ctor initializer);
CityData ctor init list reordered to declaration order; SDL3 bool
returns !-checked in blitters (dead <0 error paths now fire).

Real gameplay fixes surfaced: controlpanelwindow lost case
CP_TAB_UNIT in d7bd2561 (UnitPanelActivated unreachable) — restored;
aui_pixel uint8 loop counter could never terminate (infinite loop);
ArmyData ConvertCity wonder-bit precedence; nproposalevent no-op
statement meant to return.

Verified: full -fsyntax-only census over all 4137 TUs (both configs)
shows zero warnings; fast tests 290/290, unit 3/3, cli_save_resume
PASS on both builds.

+++ ctp2_code/gfx/spritesys/spritefile.cpp
@@ -782,8 +782,8 @@ SpriteFile::Write_v20(UnitSpriteGroup *s)
 					WriteFacedSpriteData((FacedSprite *)sprite);
 					break;
 			default:
-					c3errors_ErrorDialog("SpriteFile", "\"%s\": Bad Sprite Type %d at index %d.",
-										 m_filename,sprite->GetType(),i);
+				c3errors_ErrorDialog("SpriteFile", "\"%s\": Bad Sprite Type %d at index %zu.",
+									 m_filename,sprite->GetType(),i);
 			}
 
 			WriteAnimData(s->GetGroupAnim((UNITACTION)i));
@@ -1120,7 +1120,10 @@ void SpriteFile::ReadUnitGroup(UnitSpriteGroup *group, bool basic, int selected)
     RequireSprite(selected >= -1 && selected < ACTION_MAX, "invalid unit action index");
     bool legacy = m_version == k_SPRITEFILE_VERSION0;
     sint32 offsets[ACTION_MAX + 1];
-    int count = legacy ? UNITACTION_MAX : ACTION_MAX;
+    // UNITACTION_MAX (legacy v0 files, 5 unit actions) vs ACTION_MAX
+    // (GAME_ACTION count, current format) are distinct enums; the sprite
+    // file just stores a raw action count, so compare as plain int.
+    int count = legacy ? static_cast<int>(UNITACTION_MAX) : static_cast<int>(ACTION_MAX);
     ReadData(offsets, sizeof(sint32) * (legacy ? count : count + 1));
     if (selected == -1) {
         group->DeallocateStorage();
@@ -1226,7 +1229,10 @@ catch (std::exception const &error) {
 
 SPRITEFILEERR SpriteFile::ReadIndexed(GoodSpriteGroup *s, GAME_ACTION action) try
 {
-    RequireSprite(action >= 0 && action < GOODACTION_MAX, "invalid goods action index");
+    // action is GAME_ACTION (shared sprite-file action index) while
+    // GOODACTION_MAX bounds the goods enum; the file format intentionally
+    // reuses one action index space, so bound-check as plain int.
+    RequireSprite(action >= 0 && action < static_cast<int>(GOODACTION_MAX), "invalid goods action index");
     ReadGoodGroup(s, false);
     return SPRITEFILEERR_OK;
 }
